Create more descriptive PR names (#28)
* Create more descriptive PR names
PRs created with Maestro currently have "Update dependency files" which isn't all that useful for anyone.  This extends the title with the repos that are involved, up to a max of 80 characters, at which point it simply lists the number of input subscriptions that are involved. This also fixes an issue where the listed input builds in the PR description were listed by build asset registry database ID, rather than the build number.