Add general purpose retry and retry for a couple cases (#286)
* Add general purpose retry and retry for a couple cases
Adds a general purpose retry method (same as one from Helix API).
Use this in a couple cases to test it out and see whether the pattern works. We have seen quite a few 500 errors coming from github, and this should help. If it works I'll widen the usage out to the other Octokit calls