Shortcut dependency graph build during coherency updates (#293)
* Shortcut dependency graph build during coherency updates
We don't need to build the full graph during coherency updates each time.  Most of the time the nodes we are looking for are close to the root node of the graph.  Add the ability to break the graph build when a specified set of dependencies or assets are seen (dependency break is for common child coherency). Overall this drastically improves performance, maybe by an order of magnitude in local scenarios