[PKPIs] Product dependency flow cycles health (#436)
* [PKPIs] Subscription health metric
Implements a subscription health metric and way to query it in darc. The subscription health metric is branch+repo focused and checks the following things:
- Are there two or more subscriptions which produced the same dependency on the last build (error if yes)
- Do all input dependencies have subscriptions (error if not)
- Will these subscriptions flow automatically (warn if not)
- Are there subscriptions that do nothing (warn if there are)

Darc now implements a general get-health operation which will eventually query all of the PKPIs. Different PKPIs will apply in different ways to the system, so the get-health operation implements two general parameters which can narrow the focus of the PKPIs to:
- Specific channel (e.g. all subscriptions on dev)
- Specific repo