Replace nullable boolean logic with enums (#540)
* Replace nullable boolean logic with enums
We used nullable booleans as tri-state enums in certain cases when synchronizing pull requests. The meanings of the true, false, and null states varied, but generally did not follow a pattern of "yes", "no" and "I don't know". Unfortunately, there was also a 4th state. If SynchronizePullRequestAsync threw, we would return null (default value for nullable bool) out of ActionRunner, causing us to lose track of the PR. Exceptions are relatively rare, but the dominant exception (Octokit credentials exception) is transient and we would end up tracking the PR properly upon retry 5 minutes later.

Fix this by changing the results of SynchronizePullRequestAsync and CheckMergePolicyAsync over to enums. The default value of the enum (Invalid) will now avoid unregistering the reminder.