Implement basic caching of GitHub tree items (#582)
* Implement basic caching of GitHub tree items

We are hitting GitHub rate limits in Maestro, typically during the morning Arcade update. While there are quite a few places where we are inefficient with the rate limit, a primary one is that we are requesting the arcade common files multiple times, even though we are garuanteed to get the same file each time around.

Implement a caching system for these files. I am using the built in caching integrated into ASPNet, passing that into the darc library, and using it if available. If not available (e.g. darc client) then we don't use the caching. The cache is capped at 64MB.