Fix memory caching (again) (#588)
* Revert "Fix cache entry size setup (#587)"

This reverts commit e90b9ae6656364c5517ae20d9f5bccb0d96b2638.

* Fix memory caching (again)
When setting up the memory cache with AddMemoryCache(), the memory cache ends up shared between components. As it turns out, EF does not implement their use of the cache as a Size based cache. Thus, the service will fail in unexpected ways when performing DB queries. Instead, a custom memory cache singleton should be created so that the cache is not shared.