Skip releasing if build is in different org/project (#606)
Up to this point, we've not had a repo that had "PublishUsingPipelines" set to true and was in the DevDiv account. This doesn't work, since the release pipeline is in dnceng. This is a corner case that is not interesting because stages release handles this case already.  Skip releases if they are in a different org/project.