Process all items in RP queue (#625)
Instead of processing one item at a time, process all items in the RP queue each pass.
The previous behavior was fine until we ended up with lots of builds on stages. This caused a lot of requeuing, as builds were not yet finished when they went through the RP runner the first time around. This caused the queue length to increase, and significant delays starting the release pipelines