Use a temporary file when downloading in gather-drop (#644)
When I increased the amount of parallelism in gather-drop, I found a few cases where we may have been getting throttled and the downloads halted. When --skip-existing is passed, this means a subsequent download might not download a file which was half downloaded from the previous run. Instead of downloading directly to the target file, use a temporary file and move to the target file.