Fix location of <clear /> element (#659)
* Fix location of <clear /> element
The location of the clear element was being moved to after the maestro managed feeds. The tests passed as expected. It turns out that the xml documents are read with "preserve whitespace" on inside Maestro, but the tests did not read them as such. This triggered a bug in the feed update code, it incorrectly assumed the first element after the package sources node would be the clear element, while it was actually a whitespace node.

Fix is twofold:
- Run the tests by reading the xml documents the same way that Maestro does (use GitFileManager)
- Always remove the clear node and add it directly after the package sources node.