Add dependency flow event table (#649)
* Add DependencyFlowEvents table

This change adds the DependencyFlowEvent class and the DependencyFlowEvents table. We add each new created, completed, and updated event for every PR. These events tell us when a dependency update flowed successfully (completed and automatically merged), flowed unsuccessfully (manually merged or closed, pr updated), and when the PR was created (which will be useful when we move to the branch flow of dependency flow, in which case we will be able to say that a PR was created because the branch flow failed). We add to the table in SubscriptionActor, which should be usable by both dependency flow methods.

* Remove changes to tests

* Add tests

* Address feedback

* Use enums for DependencyFlowEvent types and reasons

* Address feedback

* Include BuildId in the DependencyFlowUpdatesTable

Also adds a "Fired" event.

* Address feedback

* Add pr url to the dependency flow event table

* Only mark dependency updates that were actually updated as failures.

* Update comments