Add dependency cycle check to barviz (#680)
* Add dependency cylce check to barviz

This change adds functionality to barviz to check the dependency graph
for dependency cycles. For each node, we walk through the dependency
tree for that node and check to see if any of its product subdependencies
share the same repository url as the node's repository url, indicating a
cycle. We record that there is a cycle and the cycle path in the
BuildData.

In the UI, we display a redo symbol to indicate that there is a cycle at
a particular node, and the hover text includes the cycle path.

* Move the cycle symbol to its own column

* Address feedback

* Address Feedback