Introduce "released" bit for builds (#698)
* Introduce "released" bit for builds
One issue with the current infra is that old builds will continue to be referenced throughout the graph because we do not ship all packages for every servicing release. This means that darc will gather the drop and include old builds. This bit gives us an easy way to trim those builds. As they are officially released, we can mark them as such and then gather-drop can trim them away automatically.

This change is the database and API change required for the feature. After this is checked in, I'll add the darc functionality and tests (using darc)