Add TimeToInclusion to BAR and BarViz
We want to know, at each level of the dependency graph, how long it took
for a dependency to get to core-sdk. So, when a change is checked into
coreclr, how long does it take to get from coreclr, through the longest
path in the dependency graph, to core-sdk. This work adds
TimeToInclusion to the BuildDependencies table. TimeToInclusion measures
how long it took from when a build was added to the Build table until
the first build that depended on it was added to the BuildDependencies
table. We then can use the TimeToInclusion in BarViz to calculate how
long it took for that dependency to appear in the dependency graph of
each repo that has it in its dependency graph. This is most important at
the core-sdk level, where we can then discover how long it takes for a
runtime change to flow all the way to the final product. This is done by
adding the time to inclusion at each level, from the root to the leafs.
So if core-setup depends directly on corefx, and flowing the corefx
build took 203 minutes, and corefx depends directly on coreclr, which
took 400 minutes to flow into corefx, then the time to inclusion for the
coreclr build into the core-setup build would be 603 minutes, which will
be reflected in BarViz.