Trigger subscription on create/update (#723)
* Prompt the user when branches or repos do not exist
One complaint from some users is that it's easy to mess up default channels, subscriptions, etc. with typos. You can easily type a wrong repo name or branch name and then be confused as to why things are not updating automatically.
To mitigate this, I've introduced checks on operations that deal with branches and repositories. These checks will make sure that branches and repositories exist, and if they do not, they will prompt the user to continue or abort.

This commit also fixes a case where AzDO will return bad request instead of 404 when checking for a branch that doesn't exist.

* Trigger subscription on create/update
When a subscription is updated or created, prompt the user to trigger the subscription (if something has changed in update).  Also add options to suppress the prompt or trigger without prompt.

* Test Fixes