Record longest build path for a build in Kusto
This change adds an algorithm that uses the TimeToInclusion metric in BAR in conjunction with build start and end time in the TimelineBuilds Kusto table to determine which path from the root node to each of its leaf nodes has the longest build time. As Kusto is the only place where we have any information on the runtime of a particular build, we need to calculate this information when we ingest into Kusto. This algorithm walks the dependency graph, stopping when we find a node that either a) has no product dependencies (a leaf) or b) finds a repository it has already seen before (a cycle). For the end node (leaf or cycle found), we use the build time by subtracting the start time from the finish time of the azdo run. For all other nodes in the path, we use the TimeToInclusionInMinutes. We do this for each build being ingested into Kusto, and record the buildid, the path that we discovered as the longest, and the time of that path. This data will be used in a power bi report to track our longest build path time over time.