Build channel removal (#788)
* Build channel removal
Adds the ability to remove builds from channels. There are times when builds get accidentally added to a channel.
In these cases, we'd like to NOT do another build, and simply roll back. This is fairly straightforward.
The build+channel combo is removed from the BuildChannels channels table, and if it happened to be that
the build was the last applied build, then re-triggering the subscription will roll back to previous build.

Added scenario tests that check for this behavior.