Preserve managed feeds outside the comment block (#789)
In some rare cases, repository owners want to explicitly place a maestro managed feed into the nuget.config file, and for various reasons, cannot directly reference a dependency that would bring that in in the version.details.xml file. In this specific case, it is because core-sdk needs to redist 3 different versions of the aspnet templates, all of which have the same package name (this is not the case for all templates, some are suffixed with major.minor).

Representing these cases in version.details.xml is a wider issue detailed in #3978. For now, though it's probably best to leave repository owners a workaround by not automatically removing maestro feeds when they are outside the managed feed block.

This change preserves those comments. Tests have been updated.