Fix CPD incorrect choice issue Up to this point, CPD has some odd quirks. It chooses the dependency that it wants to update to based how shallow the asset it finds it in the graph (lowest wins). But it does so using incomplete information in some cases. In order to improve the overall performance of CPD, most of which is graph build time, the graph build was aborted when a specific set of assets was found. This might lead to some odd choices. For instance, if a repo depends on two input dependencies from repo A, one pinned and one not pinned, then a CPD that is reachable from those two dependencies will could potentially choose to update based on the pinned input.
These choices are not wrong, but they aren't usually what the user wants. The user often just wants the path to the newest potential dependency. Doing this requires that a full build of the graph be done, which is pretty inefficient. A full build with necessary information for CPD from dotnet/extensions, for instance, was just under 2 minutes. After looking into this, I discovered that the GetBuildsAsync call was the majority of the time. Without this, the build is ~10 seconds. Most of this is because the GetBuildAsync call blocks further evaluation of the graph. On the plus side, the results of the calls aren't really needed during build, though steps need to be taken to avoid excessive calls.

What I ended up doing was starting the GetBuildsAsync calls each time a new required one was identified, then only analyzing the results at the very end of the graph build, before handing the data back to the caller.  Thus, the calls end up largely running in parallel and we only block on the results at the very end. In the process, I was able to remove a bunch of complexity as we no longer have a need for the early break build modes. The build takes at least an order of magnitude less time, quite possibly a lot more. The reuslts for a full graph build are identical to before.

The CPD algorithm is then changes to choose the newest asset with a matching name in the graph, and some additional complexity can be removed where we needed to rebuild the graph in some cases if the previous build was incomplete.