Fix using d3 with angular
Angular's minification of d3 removes functions that are used, causing a runtime error. This change brings in the d3 types to fix dagre-d3's import, adds a patch to build-angular to exclude d3 from minification, and pulls in d3.min.js from the cdn. This fixes #4758 without having to turn off all optimizations.