gather-drop improvements (#1015)
* gather-drop improvements
Contains a variety of improvements to gather-drop that focus on how the tool is normally used. The typical scenario is that someone runs gather drop to get a unified drop, the again to get the release layout. This is expensive, order dependent, and clunky. Instead, this commit changes the tool to gather both drops at the same time. In addition:
- Always use json for the manifest file (the parameter is not required)
- Call out a few commits in repos that do not have builds (because they pre-date maestro). Ignore these when determining whether there are dependencies missing builds. This is so we can use --continue-on-error.
- Trim the name, commit, and version parameters in version details files. This was causing the previous improvement to not work for a specific node.
- Copy files to both the release layout and the unified layout in one pass. This reduces the number of downloads.
- Always generate both layouts
- Remove the release layout parameter