Fix lifetime of BackgroundQueue
The BackgroundQueue took a callback, which caused almost every caller
to pass a lambda that contained a bunch of bound objects with
unpredictable lifetimes.  Since the background queue runs
asynchronously, this means that often the items complete after
the caller has returns and disposed all the bound objects,
resulting in unpredictable behavior.

Change the contract of the queue so it's no longer to
flow lifetimes into the queue.

Make the BackgroundQueue create a scope (so that various items aren't
sharing the global scoped DbContext) for each item to manage
the lifetimes appropriately, and to resolve any dependencies.