Target the solution instead of the csproj for tests
This allows the tests to be run in parallel, which significantly increases
the throughput of the tests (instead of 5 minutes, it's closer to 20 seconds).

The "projects" value is expaned by Azure DevOps and the call is made
one project at a time.  "dotnet test" has parallelism built in, but when
targeting a single project, it can't do much.