Update DarcBot to use AzureFunctionsVersion v3
Since upgrading to netcoreapp3.1 and Microsoft.Net.Sdk.Functions 3.0.2, DarcBot has been failing with the following error:

```
Cannot bind parameter 'log' to type ILogger. Make sure the parameter Type is supported by the binding.
```

This is due to not updating AzureFunctionsVersion in DarcBot from v2 to v3. This change fixes that.