Always use the auth path for downloading from public feeds (#1318)
Previously gather-drop erroneously assumed that if the project element of
the azure devops feed uri was missing, it was a private feed. In fact, it's
simply that if that element is missing, the feed is an organization scoped
feed. I believe it's only possible for a project scoped feed to be public,
so it was an assumption that worked until we created the dotnet5-internal feeds
as scoped to the internal project.

This PR contains the following:
- Rename the parameters to make more sense
- Add comments about this
- Always use the auth path to download packages.

It's valid to always use the auth path to download public packages. This doesn't
really negatively affect any user scenarios because we basically always need
the azure devops token to build the repo dependency graph, as well as a BAR token, etc.