Removing a dependency on AutoFixture
We just normalized our test dependency matrix,
and we are trying to avoid magic code.

Even after reading the documentation for AutoFixture,
I still have no idea what it's doing, and it make the test
hard to reason over (why are these things supposed to be different?
Where did they come from?)

We want to reduce the amount of libraries someone has to go
read about just to code review a simple test.