Fix HttpRequest error handling in arcade-services (#1538)
* Fix 404 error handling in GitHubClient

The message for 404 errors changed on github's side. To be safe, I've updated all of the error checking to not assume (Not Found) will be in the message.

When we move to 5.0, the HttpRequestException object also includes a StatusCode that we can reference, so we should check that instead of the message for additional safety.

* Also update feedcleaner

* Missed a using

* And a paren

* Also fix additional 401s