Fix DependencyUpdateErrorProcessor when found issue is closed
Before, we wouldn't create a new issue if there was an issue already created for the particular repo/branch combo or subscription id, whether or not it was open or closed. In the closed case, it mean no issue was logged. This change fixes the fall through so that if the issue is closed, we create a new issue. It also adds additional logging for that path.