Add Tags option and link builds and issues in build monitor
For Validate-DotNet notifications, we want to be able to distinguish builds via how the builds were tagged, and we want to be able to open issues in multiple different repos. This change adds the optional field Tags to BuildDefinition to allow us to specify tags that we would like to match on. It also turns Issues into an array, so that we can have multiple repos that we want to create issues in, and only opens the issues for a build in the linked issue via the Id field.