Fix TestData generator for nullable/global (#1698)
If there is a naming conflict in the namespace
of a test (say because the Test is in the A.B.C namespace,
and it's talking about something in the B.D namespace),
we need to add "global::" to the types to avoid the confusion.

Nullable parameters we not detected correctly either,
and resulted in "Nullable<Nullable<T>>" in the syntax.

Also, the settings file validator test didn't pass locally.