Fix feed patterns (#1753)
When the publishing change was made to move newly created feeds into the internal project vs. the organization, the regex's needed to be updated in Maestro to correctly identify this new feed pattern. The original feed pattern was based on the incorrect understanding that public feeds always had 'public' in their name, but non-public feeds never did. It was mainly confusion around project vs. org feeds.

Right now this is causing some major annoyance with the 5.0 and 3.1 internal builds. Does not affect 6.0.