Fix gather-drop's IsBlobFeedUrl for dotnetbuilds locations
IsBlobFeedUrl was used as a check to see if a location is a sleet feed, but was also repurposed to check if it was simply a blob storage location. We don't actually care if index.json is at the end of the uri in the later case, and, in fact, the dotnetbuilds locations no longer include index.json in the asset locations.

This change removes the index.json check, and adds some additional logic around removing the index.json part of the asset location uri, so that gather-drop works with dotnetbuilds locations.