Fix exceptions in TeamMentionForwarder
Github will send comment bodies as null if there is no comment body (as opposed to empty string). TeamMentionForwarder's HandleMentions assumes that the comment body will, at a minimum, be an empty string before checking to see if it contains the team name, which it uses to inform the team if there are any mentions we need to be aware of. When the comment body is null, this will throw an exception. This change adds a null check, and if the new body is null, or it doesn't contain the team name, we will not send a notification.