Make sure the MappedConfig thing logs
Since it's using System.Threading.Timer, any exception
in the callback terminates the process from a background thread,
avoiding ALL other logging.

As such, we need to log this as well.  If we need this sort of thing in more places,
we might need some sort of "ITerminalExceptionReporter" or something.