Add a null check to ProcessNotifications for edited events
When an issue is edited, it may have been edited from a null issue body (ie issueEvent.Changes.Body is null), but we assume that it is never null and blindly grab issueEvent.Changes.Body.From. This causes a lot of errors in DotNetStatus-Eng. This change adds a null check so that we don't fail in this instance.