Refactor SynchronizePackageProps to also recognize Directory.Packages.props
We have updated one of our repos to use Directory.Packages.props, but we would like to continue to use this tool. This change updates SynchronizePackageProps to also know about Directory.Packages.props. Since the format of Directory.Packages.props is also different (it uses PackageVersion instead of PackageReference), updates were also made in the parsing of the xml doc to make sure that all cases are covered.