Compile the `.gitmodules` file on every VMR commit (#1996)
The `.gitmodules` file, that has information about git submodules in a repository, is only relevant in the root of the repository. This means that we need to gather all `.gitmodules` files from all individual repos that we are synchronizing, compile a single file with all of them and place it in the root of the VMR.

This needs to happen at every commit. Additionally, we need to modify the paths to point into the `src/[repo]/..` paths. 

Also splits `IVmrManager` into `IVmrInitializer` and `IVmrUpdater` files and removes a forgotten `ignoreWorkingTree` parameter.