Add models for `git-info` files (#2010)
The tarball and the VMR contain so-called git-info files which are small XMLs where we store what revision of each individual repository the VMR was built from.
There is also another file called `AllRepoVersions.prop` which contains a list of all of the SHAs / versions extracted for easier manipulation.

The VMR tooling also needs to create/read these so that it can initialize/update the VMR appropriately. This change will come as a follow-up. This PR only adds the models and the XML de/serialization.