Reference BarViz from Api directly (#4431)
Instead of building the front-end separately and copying the static
files to the Api's `wwwroot` folder during build/publishing/docker
build, we reference the project directly and it can get the static files
automatically.
This simplifies everything for us as we can just F5 the `AppHost`
project and everything works.