Only use `maestro-auth-test` repos locally (#4438)
I discovered that using the repro tool locally does not entirely work
because during code flows, we try to use all of the previous remotes
(such as the URI stored in the `source-manifest.json`).
We cannot override those in the repositories because some scenarios such
as previous flow reconstructions would still break.

So it's easier to override the URI like this and just keep cloning the
repos that we have access to.