Query PR status only once during dependency flow (#4547)
We have 2 different calls to get PR status but both lead to the same API
endpoint (both AzDO and GitHub):
- `GetPullRequestAsync`
- `GetPullRequestStatusAsync`

We call both during dependency updates which means extra calls to
GitHub.
This PR unifies them into a single call and reuses the data loaded with
the first call instead of the second call.


https://github.com/dotnet/arcade-services/issues/4524#issuecomment-2710525969