Slightly refactor IsDevelopment check for SwaggerUI usage (#4603)
Ref: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/2222295
(MSFT internal access only)

The "only use SwaggerUI in dev" static analyzer is a bit simplistic and
only understands code of the form:

```cs
if (IsDevelopment()) {
    // enable SwaggerUI
}
```

Updating the rule is a pain, so the path of least resistance is to avoid
the early-exit from this method and instead to surround the entire
method body in an `if` block.

Note to reviewers: suggest toggling "ignore whitespace" while looking at
the diff.

---------

Co-authored-by: Přemek Vysoký <premek.vysoky@microsoft.com>