Adds a BarViz page with codeflow subscriptions  (#4606)
I polished an old branch I had going which adds a page that lists all source-enabled subscriptions. It will be useful as it we can easily see how much behind product repos / VMR is, it has links to quickly navigate
to repo CIs etc.
I will most likely use it for the presentation on Wednesday so checking it in.


https://github.com/dotnet/arcade-services/issues/3463

---------

Co-authored-by: Oleksandr Didyk <106967057+oleksandr-didyk@users.noreply.github.com>