Ignore additional mappings during codeflow (#4647)
I originally thought that we would just remove the additional mappings
from VMR's `source-mappings.json` but that does not work because the
flow will be based on the last flown commit where it's likely the
additional mappings will still be present.

So we need to explicitly disable this feature for code flows.

https://github.com/dotnet/arcade-services/issues/4618