Fix handling submodules during forward flow (#4698)
https://github.com/dotnet/arcade-services/issues/4644

There was a bug in forwardflow when the previous flow was backwards
Basically, we were leaving the submodules as is in our repo, but
creating the patches for them from 0 commit. When these patches we're
applied, we were getting exceptions that the files already exist in the
index.
This PR fixes the issue by making sure we remove the submodules at the
beginning of the forward flow, so we restore them like the normal repo,
from the 0 commit