Unify back and forward flow merge policies (#4691)
<!-- Link the GitHub or AzDO issue this pull request is associated with.
Please copy and paste the full URL rather than using the
dotnet/arcade-services# syntax -->
Previously, we had two separate classes for `ForwardFlowMergePolicy` and
`BackwardFlowMergePolicy`, and they both had the same MergePolicyName:
`Codeflow consistency check`.
Because of this, we had exceptions when we we're registering both here:
https://github.com/dotnet/arcade-services/blob/fdb14b5b71c672bb680aae643b3b5ece8483eb85/src/ProductConstructionService/ProductConstructionService.DependencyFlow/MergePolicyEvaluator.cs#L28

I personally think that it's a good idea to have these have the same
name. That way we can be sure that a subscription will only have one
codeflow merge policy, and it will be the correct one, there will be no
need for validation

https://github.com/dotnet/arcade-services/issues/4612

test: https://github.com/maestro-auth-test/maestro-test-vmr/pull/1353

---------

Co-authored-by: Přemek Vysoký <premek.vysoky@microsoft.com>