Base codeflow branches on top of last synchronization point (#4793)
Addressed problems seen in
https://github.com/dotnet/arcade-services/issues/4786 where we were
overwriting files during opposite direction flows.
This was caused by the fact that we didn't base the PR branch on top of
the last synchronized commit but rather on tip of the target branch
which was wrong. When we then removed all the files and inserted them
again, we brought back deleted files.

---------

Co-authored-by: dotnet-maestro[bot] <42748379+dotnet-maestro[bot]@users.noreply.github.com>