Silence failures from test PRs (#4895)
During tests, we delete subscriptions while the service is still
tracking open PRs.
This ends up in the background work item to end with a failure. Since
the INT environment almost exclusively triggers for tests only, they
make a majority of work items and trigger the alert for failed work
items: https://github.com/dotnet/dnceng/issues/5769

This PR makes is so that these are considered successful.