Do not keep bad policies in the policy results
Previously, any merge policy that had a cached result would be kept in the
newly evaluated results. This led to a bug where a rename to a merge policy
made us keep cached results with the old name. This commit makes it so that
only merge policies that are in the list of policy definitions are kept
in the new results.