Increase Redis operation timeout to 10s (#5011)
The most frequent exception in the service is RedisTimeout and based on
https://stackexchange.github.io/StackExchange.Redis/Timeouts we can
increase the timeout for operations that can take longer when the
instance is under load.

I checked Azure monitoring for Redis and we do cause a high amount of
reads during some spikes when these exceptions occur though we only go
to 45% load at most. So maybe this will help.

<!-- https://github.com/dotnet/arcade-services/issues/5011 -->