Fix detecting meaningless PRs (#5014)
There was a bug where we didn't correctly work with the `src/repo`
prefixes.
There was also a bug in the tests where we had the assertion
`BeTrue`/`BeFalse` swapped.
There was also a bug that the test diff had files it wouldn't have.

https://github.com/dotnet/arcade-services/issues/4178

Example PR that shouldn't have gotten opened:
https://github.com/dotnet/dotnet/pull/1331