Reconstruct previous flows iteratively until conflicts are cleared (#5086)
In https://github.com/dotnet/arcade-services/issues/5032, we found out
that if we make a series of flows and the last one has a conflict with
one that is not the most recent, we need to correctly reconstruct all
the previous ones until we find the one that is conflicting with us.
Then we base the PR branch before it and let users handle the config.

This PR makes the previous flow reconstructions algorithm work
iteratively back until it finds the one where we can apply changes on
top.