Assure requested branch exists locally in a clone (#5171)
Fixes a problem where when an already existing PR had existed, we'd
never check out the `targetBranch` of a clone.
Later, when we were trying to update its `global.json` and other files,
we were not able to retrieve the files from that branch as it didn't
exist locally.

This PR makes sure the desired branch is fetched locally in our clone
during clone preparation.

#5098